+++ spec/acceptance/daemons_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe 'phabricator::daemons' do
     Class['apt::update'] -> Package <| title != 'apt-transport-https' and title != 'ca-certificates' and title != 'software-properties-common' |>
 
     class { 'php::globals':
-      php_version => '7.1',
+      php_version => '7.2',
     }
 
     class { 'php':